First approach to InMemory Torrent storage
The problem being that either in file or BoltDB storages (a) touch the FS and
(b) it's not a easy way to delete the files after being downloaded.

Ideally, after each piece is received and processed it could be removed from the in
memory storage. This is not possible since the interfaces exposed by the torrent
library do not allow for a Close() / Delete() method on the piece.

What we've done instead is to remove all the pieces from a torrent when all pieces have
been processed. That's less than ideal, but we cannot do any better for now.