Goroutine leaking? That's the problem right here
Every time an event from the subscription was sent, a new subscription was created. That why goroutines grow that much, and the process got so slow.
Then every new event had to be pushed to 10K gorutines, blahblah. Two mistakes: create more than one subscription. Not realizing that the subscription needs to be closed.

Both realizations would have solved the problem.