Fixes bug. I did not know why events were being published as commands. Now it's obvious, I was using the very same Publisher/Subscriber.
The problem was then that the events (published as commands) were being sent to a Queue, not the exchange. It was confusing to me, and since the watermill implementation makes various uses of queue/exchange name/route keys, depending on the context, it added an extra layer of confusion. Now I know a lot more about Watermill implementation and the problem has been fixed in my end, where had been all along.