Adds coverage and fixes bug of (surprise surprise) untested code
[standing ovation] [one developer in the crowd starts crying, surprised] [irony fades out]

Moment of clarity: since I, as a programmer, and as human being, I'm unreliable, make mistakes and some of them might be unnoticed, I'm just as good as the tests are.
I like testing, it gives credability to my work, a future reference, and a safety net that makes everything easier and safer. Test are a good thing. Do tests, kids.